Commented out non-essential analyses
Commented out analyses which did not make it into the final publication but are a part of notebooks that generate the data for that publication. This way, an interested third party can run these notebooks without modification, even if they don't have those data. The omitted analyses include any involving the ETS/GABPA transcription factors or the NHF1 MNase map.